data_connectors/requirements.txt
python-dotenv==1.0.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.117.0
pypdf==5.3.1
slack-sdk==3.34.0
openai==1.66.3
nearai==0.1.16
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355